update with DMD, MRDMD, refactoring, Readme
