+++ requirements.txt
@@ -9,6 +9,6 @@ ipython
 ipywidgets
 matplotlib
 plotly
-kaleido  
+kaleido==0.1.0.post1
 imageio[ffmpeg]
 bokeh
